DIMS
Dormitory Inspection and Management System (DIMS) was developed in order to help the University to build a more suitable and functional system to manage residences.
DIMS is comprised of two main systems: staff and student. The system offers local data storage connecting with the database management system inside the University. DIMS provides an easier way for staff to set inspection times and results, and a two-way message sending function to improve communication between staff and students.